+++ src/components/Create/components/pages/PayoutsPage/index.ts
@@ -1,1 +1,1 @@
-export * from './PayoutsPage'
+export * from './OldPayoutsPage'
